No depedency to System.CodeDom when targetting .net 8.0
- CodeDom complilation cannot target .net core
- BatchCompiler.cs contains the code to complile with codedom and/or roslyn
