+++ src/Spark.JsTests/Generate.ashx.cs
@@ -1,3 +1,4 @@
+
 // Copyright 2008-2009 Louis DeJardin - http://whereslou.com
 // 
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -17,7 +18,7 @@ using System.IO;
 using System.Web;
 using System.Web.Services;
 using Spark.Bindings;
-using Spark.Compiler.Roslyn;
+using Spark.Compiler;
 using Spark.FileSystem;
 using Spark.Parser;
 using Spark.Parser.Syntax;
