+++ src/Spark.Web.Mvc/SparkView.cs
@@ -25,7 +25,7 @@ namespace Spark.Web.Mvc
         private string _siteRoot;
         private ViewDataDictionary _viewData;
         private ViewContext _viewContext;
-    	private dynamic _viewBag;
+        private dynamic _viewBag;
         
         public TempDataDictionary TempData => ViewContext.TempData;
 
@@ -52,19 +52,19 @@ namespace Spark.Web.Mvc
             set { SetViewData(value); }
         }
 
-    	public dynamic ViewBag
-    	{
-			get
-			{
-				if( _viewBag == null )
-					SetViewBag( new DynamicViewDataDictionary(ViewData) );
-				return _viewBag;
-			}
-    	}
+        public dynamic ViewBag
+        {
+            get
+            {
+                if( _viewBag == null )
+                    SetViewBag(new DynamicViewDataDictionary(ViewData));
+                return _viewBag;
+            }
+        }
 
         public ViewContext ViewContext
         {
-            get { return _viewContext; }
+            get => _viewContext;
             set { SetViewContext(value); }
         }
 
@@ -73,10 +73,10 @@ namespace Spark.Web.Mvc
             _viewData = viewData;
         }
 
-		protected virtual void SetViewBag( DynamicViewDataDictionary viewBag )
-		{
-			_viewBag = viewBag;
-		}
+        protected virtual void SetViewBag(DynamicViewDataDictionary viewBag)
+        {
+            _viewBag = viewBag;
+        }
 
         protected virtual void SetViewContext(ViewContext viewContext)
         {
@@ -134,7 +134,6 @@ namespace Spark.Web.Mvc
         {
             return ResourcePathManager.GetResourcePath(SiteRoot, path);
         }
-
 
         public void Render(ViewContext viewContext, TextWriter writer)
         {
