--- src/Spark/Compiler/CodeDom/CodeDomCompilerException.cs
@@ -1,9 +1,0 @@
-using System.CodeDom.Compiler;
-
-namespace Spark.Compiler.CodeDom
-{
-    public class CodeDomCompilerException(string message, CompilerResults results) : CompilerException(message)
-    {
-        public CompilerResults Results { get; set; } = results;
-    }
-}
